+++ include/cpprealm/internal/networking/shims.hpp
@@ -1,0 +1,41 @@
+////////////////////////////////////////////////////////////////////////////
+//
+// Copyright 2024 Realm Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+////////////////////////////////////////////////////////////////////////////
+
+#ifndef CPPREALM_SHIMS_HPP
+#define CPPREALM_SHIMS_HPP
+
+#include <cpprealm/networking/platform_networking.hpp>
+
+namespace realm {
+    namespace sync {
+        class SyncSocketProvider;
+    }
+    namespace app {
+        struct GenericNetworkTransport;
+    }
+}
+
+namespace realm::internal::networking {
+
+    std::shared_ptr<app::GenericNetworkTransport> create_http_client_shim(const std::shared_ptr<::realm::networking::http_transport_client>&);
+
+    std::unique_ptr<::realm::sync::SyncSocketProvider> create_sync_socket_provider_shim(const std::shared_ptr<::realm::networking::sync_socket_provider>& provider);
+
+}
+
+#endif //CPPREALM_SHIMS_HPP
